DirectoriesWatcher: wait the end of the file copy before launch the scan
When the event of the 'QFileSystemWatch' is received, the size of each file of
the watched directory is recorded. At the timeout event, these sizes are
compared to the current sizes in order to determine if the file is still
opened for a writing/copy.
If one file or more are still opened, the timer is restarted.
If all the files have a constant size, the scan is launched.